+++ src/components/DocumentPreviewStatus.tsx
@@ -4,7 +4,7 @@ import {EditIcon, PublishIcon} from '@sanity/icons'
 import {SanityDocument} from '@sanity/types'
 import {Box, Inline, Text, Tooltip} from '@sanity/ui'
 import React from 'react'
-import {DocumentPair} from '../types/types'
+import {DocumentPair} from '../types'
 
 export function TimeAgo({time}: {time: string | Date}) {
   const timeAgo = useTimeAgo(time)
