+++ src/components/TreeEditor.tsx
@@ -2,7 +2,7 @@ import {AddCircleIcon} from '@sanity/icons'
 import {Box, Button, Card, Flex, Spinner, Stack, Text, Tooltip} from '@sanity/ui'
 import React from 'react'
 import SortableTree from 'react-sortable-tree'
-import {SanityTreeItem, TreeInputOptions} from '../types/types'
+import {SanityTreeItem, TreeInputOptions} from '../types'
 import getCommonTreeProps, {getTreeHeight} from '../utils/getCommonTreeProps'
 import {getUnaddedItems} from '../utils/treeData'
 import useAllItems from '../utils/useAllItems'
