+++ src/parts.d.ts
