+++ src/utils/useAllItems.ts
@@ -1,7 +1,7 @@
 import {MutationEvent, SanityClient, SanityDocument} from '@sanity/client'
 import sanityClient from 'part:@sanity/base/client'
 import React from 'react'
-import {AllItems, TreeInputOptions} from '../types/types'
+import {AllItems, TreeInputOptions} from '../types'
 import {isDraft, unprefixId} from './idUtils'
 
 const client = sanityClient.withConfig({
